move att history representation into its own file
