+++ validator/db/kv/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "attestation_history.go",
         "db.go",
         "genesis.go",
+        "historical_attestations.go",
         "proposal_history.go",
         "schema.go",
     ],
@@ -30,6 +31,7 @@ go_test(
         "attestation_history_test.go",
         "db_test.go",
         "genesis_test.go",
+        "historical_attestations_test.go",
         "proposal_history_test.go",
     ],
     embed = [":go_default_library"],
